# Conflicts:
#	src/main/java/airspaceengine/routesegment/RSList.java
#	src/main/java/airspaceengine/waypoint/WPList.java

+++ src/main/java/collisionavoidanceengine/flightplan/MetaNodeRecord.java
@@ -4,7 +4,7 @@ package collisionavoidanceengine.flightplan;
  * Created by Ziji Shi on 22/12/17.
  */
 public class MetaNodeRecord {
-    public int reachTime;
+    public double reachTime;
     public String flightID;
     public boolean isLanding;
